+++ src/main/java/com/github/j5ik2o/event/store/adapter/java/internal/EventStoreAsyncForDynamoDB.java
@@ -113,8 +113,8 @@ public final class EventStoreAsyncForDynamoDB<
         null,
         null,
         new DefaultKeyResolver<>(),
-        new JsonEventSerializer<>(),
-        new JsonSnapshotSerializer<>());
+        new DefaultEventSerializer<>(),
+        new DefaultSnapshotSerializer<>());
   }
 
   @Nonnull
