+++ src/main/java/com/github/j5ik2o/event/store/adapter/java/internal/EventStoreForDynamoDB.java
@@ -112,8 +112,8 @@ public final class EventStoreForDynamoDB<
         null,
         null,
         new DefaultKeyResolver<>(),
-        new JsonEventSerializer<>(),
-        new JsonSnapshotSerializer<>());
+        new DefaultEventSerializer<>(),
+        new DefaultSnapshotSerializer<>());
   }
 
   @Nonnull
